Last Upload

+++ TiKayStore/TiKayStore/Areas/Admin/Controllers/CategoryController.cs
@@ -33,7 +33,16 @@ namespace TiKayStore.Areas.Admin.Controllers
         }
         public ActionResult Add()
         {
-            return View();
+            UserLogin user = (UserLogin)Session["USER_SESSION"];
+            var count = db.tb_Indentify.Count(x => x.IdUser == user.UserId & x.IdPermission == 9);
+            if (count > 0)
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Unpermitted", "Error");
+            }
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -61,8 +70,17 @@ namespace TiKayStore.Areas.Admin.Controllers
         }
         public ActionResult Edit(int id)
         {
-            var item = db.tb_ProductCategory.Find(id);
-            return View(item);
+            UserLogin user = (UserLogin)Session["USER_SESSION"];
+            var count = db.tb_Indentify.Count(x => x.IdUser == user.UserId & x.IdPermission == 11);
+            if (count > 0)
+            {
+                var item = db.tb_ProductCategory.Find(id);
+                return View(item);
+            }
+            else
+            {
+                return RedirectToAction("Unpermitted", "Error");
+            }
         }
         [HttpPost]
         [ValidateAntiForgeryToken]
@@ -87,14 +105,23 @@ namespace TiKayStore.Areas.Admin.Controllers
         }
         public ActionResult Delete(int id)
         {
-            var item = db.tb_ProductCategory.Find(id);
-            if (item != null)
+            UserLogin user = (UserLogin)Session["USER_SESSION"];
+            var count = db.tb_Indentify.Count(x => x.IdUser == user.UserId & x.IdPermission == 10);
+            if (count > 0)
             {
-                db.tb_ProductCategory.Remove(item);
-                db.SaveChanges();
-                return Json(new { success = true });
+                var item = db.tb_ProductCategory.Find(id);
+                if (item != null)
+                {
+                    db.tb_ProductCategory.Remove(item);
+                    db.SaveChanges();
+                    return Json(new { success = true });
+                }
+                return Json(new { success = false });
             }
-            return Json(new { success = false });
+            else
+            {
+                return Json(new { success = false });
+            }
         }
     }
 }
